requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
coverage==7.3.0
Django==4.2.4
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
docutils==0.20.1
drf-yasg==1.21.7
idna==3.4
inflection==0.5.1
packaging==23.1
Pillow==10.0.0
psycopg2-binary==2.9.7
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
sqlparse==0.4.4
stripe==6.5.0
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==2.0.5
